+++ DataAccessLayer/DataObjets/user.java
@@ -1,4 +1,4 @@
-package DataAccessLayer;
+package DataAccessLayer.DataObjets;
 
 import java.util.ArrayList;
 import java.util.Collections;
